+++ Assets/Sources/Runtime/Signal/SignalChangeHealth.cs
@@ -4,5 +4,5 @@ using Pixeye.Actors;
 public struct SignalChangeHealth
 {
 	public ent target;
-	public int count;
+	public float count;
 }
